Created Model.cs for the Model and Commands.cs for the commands
I just thought this would be a bit easier to understand than having everything in MainWindow.xaml.cs
