+++ ColorConverter-WPF/ColorConverter-WPF.csproj
@@ -65,6 +65,8 @@
     </Compile>
     <Compile Include="ColorCalculator.cs" />
     <Compile Include="ColorConverter.cs" />
+    <Compile Include="Commands.cs" />
+    <Compile Include="Converters.cs" />
     <Compile Include="InputChecker.cs" />
     <Compile Include="MainWindow.xaml.cs">
       <DependentUpon>MainWindow.xaml</DependentUpon>
@@ -72,6 +74,7 @@
     </Compile>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Model.cs" />
     <Compile Include="Properties\AssemblyInfo.cs">
       <SubType>Code</SubType>
     </Compile>
