+++ ColorConverter-WPF/Commands.cs
@@ -1,0 +1,14 @@
+﻿using System.Windows.Input;
+
+namespace ColorConverter_WPF
+{
+    public static class CustomCommands
+    {
+        public static readonly RoutedUICommand CheckInput = new RoutedUICommand
+            (
+            "CheckInput",
+            "CheckInput",
+            typeof(CustomCommands)
+            );
+    }
+}
